Homework 3: class polymorphism and ClassCastException
