+++ src/main/java/lesson_3_polymorphism/Electric_Car.java
@@ -1,6 +1,8 @@
-package lesson_1;
+package lesson_3_polymorphism;
 
-public class Electric_Car extends Car{
+import lesson_3_polymorphism.Car;
+
+public class Electric_Car extends Car {
     public Electric_Car(String name, int id, String color) {
         super(name, id, color);
     }
